requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
openpyxl>=3.1.0